240411 :  fastapi 성공! 근데 csv대신 오라클 디비 연동해서 db값과 비교해야해

+++ lstm/src/main/java/com/example/lstm/entity/ProdEntity.java
@@ -72,7 +72,7 @@ public class ProdEntity {
     private int hitCount;
 
     @Column(name = "lstm_predict_proba", nullable = false)
-    private float lstmPredictProba;
+    private double lstmPredictProba;
 
     @Column(name = "lstm_predict", nullable = false)
     private boolean lstmPredict;
